refactor(core): split @linchkit/core barrels by domain (closes #371)
CodeRabbit has flagged the 500-line file limit on every PR that adds
to the public barrels for months. Root cause: every engine + Capability
re-exports through two monolithic entry files.

- `packages/core/src/index.ts`:        495 → 47 lines (thin aggregator)
- `packages/core/src/server-entry.ts`: 547 → 29 lines (thin aggregator)

Split exports into two parallel sub-barrel trees under
`packages/core/src/exports/`:

  client/  (22 sub-barrels) — what the browser-safe `@linchkit/core`
           entry surfaces; one file per source-domain folder
  server/  (15 sub-barrels) — what `@linchkit/core/server` adds on top;
           includes persistence + addon-discovery + deployment

Largest sub-barrel: `exports/server/ai.ts` at 114 lines. All others
under 100 — plenty of headroom before any single file approaches the
500-line ceiling.

Why two parallel trees instead of one shared tree: some names exist
as type-only on the browser side and value-bearing on the server side
(e.g. `ActionRegistry`). A single shared barrel using `export *` would
silently drop colliding names due to TypeScript's collision rule. The
duplication between `client/<name>.ts` and `server/<name>.ts` is
small (~20 line items each) and the mapping is mechanical: new file
in `src/<domain>/` → re-export it from both `client/<domain>.ts` and
`server/<domain>.ts` if it belongs in both surfaces.

Four names (`ValidationResult`, `SlotName`, `ValidationError`,
`SignalBus`) live in both `types.ts` and a more specific sub-barrel.
Pre-split this was resolved by TypeScript's "named re-export beats
`export *`" rule. Restored with four explicit re-export lines at the
top of `index.ts`, with a comment block documenting why.

## Public surface preservation

New test `packages/core/__tests__/barrel-public-surface.test.ts`
snapshots every runtime export from both entries:

- `@linchkit/core`         — 161 names (VERSION + 160 classes/functions/constants)
- `@linchkit/core/server`  — 175 names

The lists are bit-for-bit identical to what `main` exposed. The full
repo `tsc --noEmit` is clean — every existing addon, test, and CLI
file compiles without modification.

Zero consumer files touched. No public API renames. No new deps.
