+++ packages/core/__tests__/barrel-public-surface.test.ts
@@ -1,0 +1,370 @@
+import { describe, expect, test } from "bun:test";
+
+/**
+ * Locks the public export surface of @linchkit/core and @linchkit/core/server.
+ *
+ * Both entry barrels were split into focused sub-barrels under
+ * src/exports/{client,server}/* to stay under the 500-line file size limit.
+ * This snapshot guards against an accidental drop or addition of an exported
+ * name during future barrel edits — every public name must appear here and
+ * every name here must still be exported.
+ *
+ * When intentionally adding/removing a name, update the corresponding array
+ * in the SAME commit as the change. Treat surface drops as breaking changes.
+ */
+
+const EXPECTED_CLIENT_EXPORTS = [
+  "AuthenticationError",
+  "AuthorizationError",
+  "BusinessRuleError",
+  "CapabilityHub",
+  "CapabilityHubError",
+  "ConfigRegistry",
+  "ConfigValidationError",
+  "ConflictError",
+  "DEFAULT_EXECUTION_META_MASKED_KEYS",
+  "DEFAULT_META_MAX_BYTES",
+  "DerivedPropertyEngine",
+  "ERROR_STATUS_MAP",
+  "I18N_RAW_KEY",
+  "InMemoryConfigStore",
+  "LinchKitError",
+  "LocalCapabilityRegistry",
+  "MERGEABLE_CONSTRAINT_KEYS",
+  "MetaSizeError",
+  "NoopMeter",
+  "NoopTracer",
+  "NotFoundError",
+  "OPTIONAL_ENV_VARS",
+  "REQUIRED_ENV_VARS",
+  "RuntimeConfigRegistry",
+  "SystemError",
+  "TRANSLATABLE_FIELD_TYPES",
+  "VERSION",
+  "ValidationError",
+  "aggregateReleaseType",
+  "analyzeFile",
+  "analyzeMigrationSql",
+  "applyOverride",
+  "buildActionChain",
+  "buildProjectOverview",
+  "buildRelationGraph",
+  "buildResult",
+  "canAutoApproveOverlayChange",
+  "canAutoApproveOverlayProposal",
+  "capabilityCategoryEnum",
+  "capabilityMetadataSchema",
+  "capabilityTypeEnum",
+  "checkReleaseCompatibility",
+  "checkTrustPermissions",
+  "classifyStatement",
+  "clearDoctorChecks",
+  "computeAggregate",
+  "createCapabilityHub",
+  "createDedupAnalyzer",
+  "createDefaultInsightTranslatorRegistry",
+  "createDerivedPropertyEngine",
+  "createDispatchQuery",
+  "createEvolutionRuntime",
+  "createExecutionMeta",
+  "createExtensionResolver",
+  "createImpactAnalyzer",
+  "createInsightTranslatorRegistry",
+  "createInterfaceRegistry",
+  "createLocalRegistry",
+  "createPreAnalysisPipeline",
+  "createRelationRegistry",
+  "createSagaRunner",
+  "createSignalBus",
+  "createTranslatableValue",
+  "databaseConfig",
+  "deepMerge",
+  "defineAction",
+  "defineCapability",
+  "defineConfig",
+  "defineConfigSchema",
+  "defineDataAccess",
+  "defineEntity",
+  "defineEvent",
+  "defineEventHandler",
+  "defineInterface",
+  "definePermissionGroup",
+  "defineRelation",
+  "defineRule",
+  "defineSaga",
+  "defineSemanticRelation",
+  "defineSensor",
+  "defineState",
+  "defineView",
+  "defineWatcher",
+  "describeAction",
+  "describeEntity",
+  "describeRelation",
+  "detectLocale",
+  "disableRule",
+  "evaluateCondition",
+  "evaluateExpression",
+  "executeOverlayProposal",
+  "executionConfig",
+  "extendEntity",
+  "extendPermissionGroup",
+  "extendState",
+  "extendView",
+  "field",
+  "filterEntityByCapabilities",
+  "findSensor",
+  "formLayout",
+  "generateAgentsMd",
+  "generateExpressionIndex",
+  "generateGinIndex",
+  "generateTranslatableIndexes",
+  "generateZodSchema",
+  "getDerivedStrategy",
+  "getDoctorChecks",
+  "getObservability",
+  "getSensors",
+  "getTranslatableFields",
+  "group",
+  "inferSemanticRelations",
+  "initI18n",
+  "isAiAgentCaller",
+  "isDerivedField",
+  "mergeFieldDefinition",
+  "mergeTranslatableValue",
+  "noopMeter",
+  "noopTracer",
+  "normalizeTranslatableRow",
+  "normalizeTranslatableValue",
+  "notebook",
+  "overrideAction",
+  "overrideEntity",
+  "overrideRule",
+  "page",
+  "parseAcceptLanguage",
+  "queueConfig",
+  "redactMetaForLog",
+  "registerDoctorCheck",
+  "registerSensor",
+  "registerTranslations",
+  "resetObservability",
+  "resolveAggregateValue",
+  "resolveDerivedValue",
+  "resolveEnvVars",
+  "resolveField",
+  "resolveLabel",
+  "resolveLocale",
+  "resolveOverrides",
+  "resolveRuleOverride",
+  "resolveTranslatableRow",
+  "resolveTranslatableValue",
+  "resolveTranslation",
+  "resolveWithCascade",
+  "row",
+  "runSaga",
+  "satisfiesVersionRange",
+  "securityConfig",
+  "separator",
+  "serverConfig",
+  "setObservability",
+  "shouldIncludeErrorContext",
+  "splitStatements",
+  "stripSystemKeys",
+  "unregisterSensor",
+  "validateCapabilityMetadata",
+  "validateEnv",
+  "validateIdentifier",
+  "validateTranslatableEntity",
+  "wrapTranslatableValue",
+] as const;
+
+const EXPECTED_SERVER_EXPORTS = [
+  "AIAuditLogger",
+  "AIBoundary",
+  "AIBoundaryError",
+  "ActionRegistry",
+  "AlertEngine",
+  "BatchValidationError",
+  "CACHE_INVALIDATION_CHANNEL",
+  "CacheManager",
+  "DEFAULT_COMPUTE_TIMEOUT_MS",
+  "DefaultOverlayRegistry",
+  "DeployWebhookHandler",
+  "DerivedPropertyEngine",
+  "DrizzleApprovalStore",
+  "DrizzleConfigStore",
+  "DrizzleDataProvider",
+  "DrizzleExecutionLogger",
+  "DrizzleOverlayStore",
+  "DrizzleTransactionManager",
+  "EntityRegistry",
+  "EventBus",
+  "EventHandlerRegistry",
+  "ExposureError",
+  "FlowRegistryImpl",
+  "GracefulShutdownManager",
+  "HealthCheckRegistry",
+  "INTENT_ALTERNATIVES_CONFIDENCE_THRESHOLD",
+  "INTENT_DEFAULT_MAX_HISTORY_MESSAGES",
+  "INTENT_MAX_ALTERNATIVES",
+  "INTENT_MIN_CONFIDENCE",
+  "InMemoryApprovalStore",
+  "InMemoryCacheProvider",
+  "InMemoryExecutionLogger",
+  "InMemoryMetricsCollector",
+  "InMemoryOverlayStore",
+  "InMemoryStore",
+  "InterfaceRegistry",
+  "MAX_BATCH_SIZE",
+  "MAX_CHAIN_DEPTH",
+  "NoopMeter",
+  "NoopTracer",
+  "OnchangeEvaluatorError",
+  "OverlayAwareDataProvider",
+  "PermissionRegistry",
+  "PersistentEventBus",
+  "PipelineError",
+  "PostgresCacheInvalidator",
+  "ProposalEngine",
+  "ProposalFileWriter",
+  "ProposalGenerationError",
+  "ProposalGitCommitter",
+  "RelationRegistry",
+  "TableRegistry",
+  "TenantOverrideStore",
+  "approvalStatusEnum",
+  "approvalsTable",
+  "buildColumn",
+  "buildCompatibilitySnapshot",
+  "buildIntentSystemPrompt",
+  "buildSystemColumns",
+  "buildTableColumns",
+  "builtinChecks",
+  "bumpVersion",
+  "canTransition",
+  "canUnmask",
+  "checkActionPermission",
+  "checkConnection",
+  "clearDoctorChecks",
+  "closeDatabase",
+  "collectRules",
+  "compatibilityCheck",
+  "configScopeEnum",
+  "configTable",
+  "configVersionsTable",
+  "consoleLogger",
+  "createActionExecutor",
+  "createApprovalEngine",
+  "createApprovalVerifier",
+  "createAttentionBudget",
+  "createAwarenessEngine",
+  "createCacheCheck",
+  "createCommandLayer",
+  "createDatabase",
+  "createDatabaseCheck",
+  "createDerivedPropertyEngine",
+  "createDispatchQuery",
+  "createDlqService",
+  "createEntityCheck",
+  "createEntityRegistry",
+  "createEventBus",
+  "createEventBusCheck",
+  "createEventReplayService",
+  "createEvolutionRuntime",
+  "createExtendedProposalValidator",
+  "createFlowRegistry",
+  "createFlowStepContext",
+  "createInterfaceRegistry",
+  "createNoopAIService",
+  "createOnchangeEvaluator",
+  "createOntologyRegistry",
+  "createOutboxWorker",
+  "createPersistentEventBus",
+  "createProposalEngine",
+  "createProposalGenerator",
+  "createProposalGitCommitter",
+  "createProposalValidator",
+  "createRelationRegistry",
+  "createSignalBus",
+  "createStateMachine",
+  "createStructuredLogger",
+  "createSyncFlowEngine",
+  "createTenantAwareDataProvider",
+  "createTenantIsolationMiddleware",
+  "createTestLogSink",
+  "createTriggerBinding",
+  "createUsageImportanceGraph",
+  "createWatcherRegistry",
+  "defaultTenantResolver",
+  "defineSensor",
+  "defineSystemAlert",
+  "detectEnvironment",
+  "detectInjection",
+  "drizzleSchema",
+  "dryRunProposal",
+  "evaluateConditions",
+  "evaluateRules",
+  "eventStatusEnum",
+  "eventsTable",
+  "executeBatch",
+  "executionStatusEnum",
+  "executionsTable",
+  "fieldOverlaysTable",
+  "generateDrizzleSchemaFile",
+  "generateDrizzleTable",
+  "generateFieldCode",
+  "generateMigrationSql",
+  "generatePromotionPlan",
+  "generateRelationColumns",
+  "getAvailableActions",
+  "getAvailableTransitions",
+  "getCurrentTrace",
+  "getDoctorChecks",
+  "getObservability",
+  "getTraceDepth",
+  "linchkitSchema",
+  "livenessCheck",
+  "maskRecord",
+  "maskRecords",
+  "maskValue",
+  "noopMeter",
+  "noopMetricsCollector",
+  "noopTracer",
+  "overlayStatusEnum",
+  "overrideTargetTypeEnum",
+  "registerDoctorCheck",
+  "resetObservability",
+  "resolveAutoInstall",
+  "resolveConditionVariables",
+  "resolveDataAccess",
+  "resolveFieldMasking",
+  "resolveIntent",
+  "sanitizeAIOutput",
+  "sanitizePII",
+  "sanitizePrompt",
+  "sanitizeRecordForAI",
+  "scanAddonsPath",
+  "setObservability",
+  "tenantOverridesTable",
+  "transition",
+  "validateAIOutput",
+  "validateAIProposal",
+  "validatePhase1",
+  "validateProposal",
+  "validateProposalExtended",
+  "validateRequiredEnvVars",
+  "withTrace",
+] as const;
+
+describe("public barrel surface", () => {
+  test("@linchkit/core exports the expected runtime names", async () => {
+    const mod = await import("@linchkit/core");
+    const actual = Object.keys(mod).sort();
+    expect(actual).toEqual([...EXPECTED_CLIENT_EXPORTS]);
+  });
+
+  test("@linchkit/core/server exports the expected runtime names", async () => {
+    const mod = await import("@linchkit/core/server");
+    const actual = Object.keys(mod).sort();
+    expect(actual).toEqual([...EXPECTED_SERVER_EXPORTS]);
+  });
+});
