+++ packages/core/src/exports/client/ai.ts
@@ -1,0 +1,17 @@
+/**
+ * AI types (browser-safe).
+ * Runtime AI classes live in ../server/ai.ts.
+ */
+
+export type {
+  AIActionAccess,
+  AIBoundaryCheckResult,
+  AIBoundaryOptions,
+  AIBudget,
+  AIBudgetConfig,
+  AICallRequest,
+  AIContentFilter,
+  AIPolicy,
+  AIRateLimits,
+  AIUsageRecord,
+} from "../../ai";
