+++ packages/core/src/exports/client/cache.ts
@@ -1,0 +1,14 @@
+/**
+ * Cache types (browser-safe).
+ * Runtime cache providers live in ../server/cache.ts.
+ */
+
+export type {
+  CacheEntry,
+  CacheManagerOptions,
+  CacheProvider,
+  CacheSetOptions,
+  CacheStats,
+  InMemoryCacheOptions,
+  NamespacedCache,
+} from "../../cache";
