+++ packages/core/src/exports/client/capability.ts
@@ -1,0 +1,34 @@
+/**
+ * Capability hub — discovery, dependency, extension resolution (browser-safe).
+ */
+
+export type {
+  ActionOverrideEntry,
+  CapabilityDependency,
+  CapabilityManifest,
+  CapabilityProvides,
+  CapabilityRequires,
+  CapabilitySearchOptions,
+  CompatibilityIssue,
+  EntityExtensionEntry,
+  EntityOverrideEntry,
+  ExtensionResolver,
+  RegistryEntry,
+  RegistrySearchOptions,
+  ResolutionConflict,
+  RuleOverrideEntry,
+  TrustLevel,
+  ValidationResult,
+} from "../../capability";
+export {
+  buildActionChain,
+  CapabilityHub,
+  CapabilityHubError,
+  checkTrustPermissions,
+  createCapabilityHub,
+  createExtensionResolver,
+  createLocalRegistry,
+  filterEntityByCapabilities,
+  LocalCapabilityRegistry,
+  satisfiesVersionRange,
+} from "../../capability";
