+++ packages/core/src/exports/client/config.ts
@@ -1,0 +1,28 @@
+/**
+ * Config center — schemas, in-memory store, declarative configs (browser-safe).
+ */
+
+export type {
+  ConfigEntry,
+  ConfigSchemaRef,
+  ConfigScope,
+  ConfigScopeRef,
+  ConfigStore,
+  ConfigValueHistoryEntry,
+  ConfigVersion,
+  SetConfigOptions,
+} from "../../config";
+export {
+  ConfigRegistry,
+  ConfigValidationError,
+  DEFAULT_EXECUTION_META_MASKED_KEYS,
+  databaseConfig,
+  defineConfigSchema,
+  executionConfig,
+  InMemoryConfigStore,
+  queueConfig,
+  RuntimeConfigRegistry,
+  resolveWithCascade,
+  securityConfig,
+  serverConfig,
+} from "../../config";
