+++ packages/core/src/exports/client/define.ts
@@ -1,0 +1,28 @@
+/**
+ * Declarative definition helpers — defineEntity, defineAction, defineRule, etc.
+ */
+
+export {
+  defineAction,
+  defineCapability,
+  defineConfig,
+  defineDataAccess,
+  defineEntity,
+  defineEvent,
+  defineEventHandler,
+  defineInterface,
+  definePermissionGroup,
+  defineRelation,
+  defineRule,
+  defineState,
+  defineView,
+  defineWatcher,
+  disableRule,
+  extendEntity,
+  extendPermissionGroup,
+  extendState,
+  extendView,
+  overrideAction,
+  overrideEntity,
+  overrideRule,
+} from "../../define";
