+++ packages/core/src/exports/client/doctor.ts
@@ -1,0 +1,12 @@
+/**
+ * Doctor — project health check registry (browser-safe; server adds builtin checks).
+ */
+
+export type {
+  CheckCategory,
+  CheckStatus,
+  DoctorCheck,
+  DoctorCheckResult,
+  DoctorContext,
+} from "../../doctor";
+export { clearDoctorChecks, getDoctorChecks, registerDoctorCheck } from "../../doctor";
