+++ packages/core/src/exports/client/errors.ts
@@ -1,0 +1,17 @@
+/**
+ * Error classes + helpers (browser-safe).
+ */
+
+export type { ToResponseOptions } from "../../errors";
+export {
+  AuthenticationError,
+  AuthorizationError,
+  BusinessRuleError,
+  ConflictError,
+  isAiAgentCaller,
+  LinchKitError,
+  NotFoundError,
+  SystemError,
+  shouldIncludeErrorContext,
+  ValidationError,
+} from "../../errors";
