+++ packages/core/src/exports/client/event.ts
@@ -1,0 +1,19 @@
+/**
+ * Event types (browser-safe).
+ * Runtime EventBus + replay/dlq services live in ../server/event.ts.
+ */
+
+export type { EventBus, EventHandlerRegistry } from "../../event/event-bus";
+export type {
+  BatchReplayResult,
+  EventDetail,
+  EventListOptions,
+  EventReplayService,
+  EventReplayServiceOptions,
+  EventSummary,
+  HandlerExecution,
+  HandlerHistoryQuery,
+  ReplayError,
+  ReplayOptions,
+  ReplayResult,
+} from "../../event/event-replay-service";
