+++ packages/core/src/exports/client/flow.ts
@@ -1,0 +1,12 @@
+/**
+ * Flow engine types (browser-safe).
+ * Runtime flow engine + trigger-binding helpers live in ../server/flow.ts.
+ */
+
+export type {
+  FlowEngine,
+  FlowRegistry,
+  FlowStepContext,
+  FlowStepContextDeps,
+  TriggerBinding,
+} from "../../flow";
