+++ packages/core/src/exports/client/i18n.ts
@@ -1,0 +1,14 @@
+/**
+ * i18n — shared types, locale utilities, core label resolver (browser-safe).
+ */
+
+export {
+  detectLocale,
+  type I18nConfig,
+  initI18n,
+  parseAcceptLanguage,
+  registerTranslations,
+  resolveLabel,
+  resolveLocale,
+  type SupportedLanguage,
+} from "../../i18n";
