+++ packages/core/src/exports/client/security.ts
@@ -1,0 +1,7 @@
+/**
+ * Security — data masking type contracts (browser-safe).
+ * Runtime masking helpers + tenant isolation live in ../server/security.ts
+ * (uses node:crypto).
+ */
+
+export type { MaskRecordOptions } from "../../security";
