+++ packages/core/src/exports/client/utils.ts
@@ -1,0 +1,7 @@
+/**
+ * Misc utilities (browser-safe).
+ */
+
+export { resolveEnvVars } from "../../utils/env";
+export type { IdentifierValidationResult } from "../../utils/identifier";
+export { validateIdentifier } from "../../utils/identifier";
