+++ packages/core/src/exports/client/view.ts
@@ -1,0 +1,15 @@
+/**
+ * View layout helpers — syntactic sugar over FormLayoutNode JSON shape.
+ * Browser-safe.
+ */
+
+export { type FormLayoutBuilder, formLayout } from "../../view/form-layout-builder";
+export {
+  type FieldOptions,
+  field,
+  group,
+  notebook,
+  page,
+  row,
+  separator,
+} from "../../view/layout-helpers";
