+++ packages/core/src/exports/server/addon-discovery.ts
@@ -1,0 +1,6 @@
+/**
+ * Addon discovery (Spec 57) — filesystem scanner + auto-install resolver.
+ */
+
+export { scanAddonsPath } from "../../capability/addon-scanner";
+export { resolveAutoInstall } from "../../capability/auto-install";
