+++ packages/core/src/exports/server/cache.ts
@@ -1,0 +1,19 @@
+/**
+ * Cache runtime — manager, in-memory + Postgres invalidator (server-only).
+ */
+
+export {
+  CACHE_INVALIDATION_CHANNEL,
+  type CacheEntry,
+  type CacheInvalidationPayload,
+  CacheManager,
+  type CacheManagerOptions,
+  type CacheProvider,
+  type CacheSetOptions,
+  type CacheStats,
+  type InMemoryCacheOptions,
+  InMemoryCacheProvider,
+  type NamespacedCache,
+  PostgresCacheInvalidator,
+  type PostgresCacheInvalidatorOptions,
+} from "../../cache";
