+++ packages/core/src/exports/server/event.ts
@@ -1,0 +1,35 @@
+/**
+ * Event bus runtime — bus, persistent bus, DLQ, replay, outbox (server-only).
+ */
+
+export {
+  createDlqService,
+  type DlqEntry,
+  type DlqListOptions,
+  type DlqService,
+  type DlqStats,
+} from "../../event/dlq-service";
+export { createEventBus, EventBus, EventHandlerRegistry } from "../../event/event-bus";
+export {
+  type BatchReplayResult,
+  createEventReplayService,
+  type EventDetail,
+  type EventListOptions,
+  type EventReplayService,
+  type EventReplayServiceOptions,
+  type EventSummary,
+  type HandlerExecution,
+  type HandlerHistoryQuery,
+  type ReplayError,
+  type ReplayOptions,
+  type ReplayResult,
+} from "../../event/event-replay-service";
+export {
+  createOutboxWorker,
+  type OutboxWorker,
+  type OutboxWorkerOptions,
+} from "../../event/outbox-worker";
+export {
+  createPersistentEventBus,
+  PersistentEventBus,
+} from "../../event/persistent-event-bus";
