+++ packages/core/src/exports/server/flow.ts
@@ -1,0 +1,17 @@
+/**
+ * Flow engine runtime (server-only).
+ */
+
+export {
+  createFlowRegistry,
+  createFlowStepContext,
+  createSyncFlowEngine,
+  createTriggerBinding,
+  type FlowEngine,
+  type FlowExecuteActionOptions,
+  type FlowRegistry,
+  FlowRegistryImpl,
+  type FlowStepContext,
+  type FlowStepContextDeps,
+  type TriggerBinding,
+} from "../../flow";
