+++ packages/core/src/exports/server/ontology.ts
@@ -1,0 +1,11 @@
+/**
+ * Ontology — unified semantic facade factory + descriptor types (server).
+ */
+
+export {
+  createOntologyRegistry,
+  type EntityDescriptor,
+  type OntologyRegistry,
+  type OntologyRegistryDeps,
+  type RelationDescriptor,
+} from "../../ontology";
