+++ packages/core/src/exports/server/security.ts
@@ -1,0 +1,19 @@
+/**
+ * Security runtime — data masking (node:crypto) + tenant isolation (server-only).
+ */
+
+export {
+  canUnmask,
+  type MaskRecordOptions,
+  maskRecord,
+  maskRecords,
+  maskValue,
+  resolveFieldMasking,
+} from "../../security";
+export {
+  createTenantAwareDataProvider,
+  createTenantIsolationMiddleware,
+  defaultTenantResolver,
+  type TenantIsolationMiddlewareOptions,
+  type TenantResolver,
+} from "../../security/tenant-isolation";
