+++ packages/core/src/server-entry.ts
@@ -5,543 +5,25 @@
  * NOT safe for browser — requires Node/Bun runtime.
  *
  * Usage: import { createActionExecutor, EntityRegistry } from "@linchkit/core/server"
+ *
+ * The barrel is composed from focused sub-barrels under ./exports/server/*.
+ * Each sub-barrel groups one domain (engines, persistence, AI, etc.) and stays
+ * well below the 500-line file size limit so this entry never has to grow as
+ * new engines and capabilities land.
  */
 
-// === Watcher registry (data-condition automation, spec 45) ===
-// WatcherEngine concrete impl moved to @linchkit/cap-ai-provider
-// (Spec 56 Phase 2 Step 2c) — the abstract `Watcher` lifecycle interface
-// lives in `./life-system/watcher.ts`; the registry stays in core because
-// the action pipeline consumes WatcherDefinition records directly.
-export { createWatcherRegistry, type WatcherRegistry } from "./automation";
-
-// === Engine: action, command, approval, state, rule, validation, permission, proposal ===
-
-export {
-  type ActionExecutor,
-  type ActionExecutorOptions,
-  ActionRegistry,
-  createActionExecutor,
-  type DataProvider,
-  type DataQueryOptions,
-  type ExecuteOptions,
-  type ExecutionChannel,
-  type PendingEvent,
-  type TransactionManager,
-} from "./engine/action-engine";
-export {
-  type ApprovalEngine,
-  type ApprovalEngineOptions,
-  type CreateApprovalOptions,
-  createApprovalEngine,
-  createApprovalVerifier,
-  InMemoryApprovalStore,
-} from "./engine/approval-engine";
-// Batch action engine (Spec 04 §8, Spec 16 §2.1)
-export {
-  BatchValidationError,
-  type ExecuteBatchOptions,
-  executeBatch,
-  MAX_BATCH_SIZE,
-} from "./engine/batch-action-engine";
-
-export {
-  type CommandBatchExecuteOptions,
-  type CommandContext,
-  type CommandExecuteOptions,
-  type CommandLayer,
-  type CommandLayerOptions,
-  createCommandLayer,
-  ExposureError,
-  type MiddlewareHandler,
-  type MiddlewareRegistration,
-  PipelineError,
-  type SlotName,
-} from "./engine/command-layer";
-export {
-  createOnchangeEvaluator,
-  DEFAULT_COMPUTE_TIMEOUT_MS,
-  MAX_CHAIN_DEPTH,
-  type OnchangeEvaluateArgs,
-  type OnchangeEvaluationResult,
-  type OnchangeEvaluator,
-  OnchangeEvaluatorError,
-  type OnchangeEvaluatorErrorCode,
-  type OnchangeEvaluatorOptions,
-  type OnchangeReadPermissionCheck,
-} from "./engine/onchange-evaluator";
-export {
-  checkActionPermission,
-  PermissionRegistry,
-  resolveConditionVariables,
-  resolveDataAccess,
-} from "./engine/permission-engine";
-export {
-  bumpVersion,
-  type CreateProposalOptions,
-  createProposalEngine,
-  type OnApprovedHook,
-  ProposalEngine,
-  type ProposalEngineOptions,
-} from "./engine/proposal-engine";
-export {
-  ProposalFileWriter,
-  type ProposalFileWriterOptions,
-} from "./engine/proposal-file-writer";
-export {
-  createProposalGenerator,
-  ProposalGenerationError,
-  type ProposalGeneratorDeps,
-} from "./engine/proposal-generator";
-export {
-  createProposalGitCommitter,
-  type ProposalGhRunner,
-  type ProposalGitCommitResult,
-  ProposalGitCommitter,
-  type ProposalGitCommitterOptions,
-  type ProposalGitCommitterRunResult,
-  type ProposalGitRunner,
-} from "./engine/proposal-git-committer";
-
-export {
-  collectRules,
-  evaluateConditions,
-  evaluateRules,
-  type RuleEvalInput,
-  type RuleEvalOptions,
-  type RuleEvalOutput,
-} from "./engine/rule-engine";
-
-export type { StateMachine } from "./engine/state-machine";
-export {
-  canTransition,
-  createStateMachine,
-  getAvailableActions,
-  getAvailableTransitions,
-  transition,
-} from "./engine/state-machine";
-
-export {
-  type ValidationContext,
-  validatePhase1,
-  validateProposal,
-} from "./engine/validation-engine";
-
-// === Schema registry ===
-
-export {
-  createDerivedPropertyEngine,
-  DerivedPropertyEngine,
-} from "./entity/derived-property";
-export { createInterfaceRegistry, InterfaceRegistry } from "./entity/entity-interface";
-export { createEntityRegistry, EntityRegistry } from "./entity/entity-registry";
-export {
-  buildColumn,
-  buildSystemColumns,
-  buildTableColumns,
-  type DrizzleGeneratorOptions,
-  generateDrizzleTable,
-  generateRelationColumns,
-  type RelationColumnsResult,
-} from "./entity/entity-to-drizzle";
-export { generateDrizzleSchemaFile } from "./entity/generate-drizzle-schema";
-export { createRelationRegistry, RelationRegistry } from "./entity/relation-registry";
-
-// === Event bus ===
-
-export {
-  createDlqService,
-  type DlqEntry,
-  type DlqListOptions,
-  type DlqService,
-  type DlqStats,
-} from "./event/dlq-service";
-export { createEventBus, EventBus, EventHandlerRegistry } from "./event/event-bus";
-export {
-  type BatchReplayResult,
-  createEventReplayService,
-  type EventDetail,
-  type EventListOptions,
-  type EventReplayService,
-  type EventReplayServiceOptions,
-  type EventSummary,
-  type HandlerExecution,
-  type HandlerHistoryQuery,
-  type ReplayError,
-  type ReplayOptions,
-  type ReplayResult,
-} from "./event/event-replay-service";
-export {
-  createOutboxWorker,
-  type OutboxWorker,
-  type OutboxWorkerOptions,
-} from "./event/outbox-worker";
-export { createPersistentEventBus, PersistentEventBus } from "./event/persistent-event-bus";
-
-// === Cache ===
-
-export {
-  CACHE_INVALIDATION_CHANNEL,
-  type CacheEntry,
-  type CacheInvalidationPayload,
-  CacheManager,
-  type CacheManagerOptions,
-  type CacheProvider,
-  type CacheSetOptions,
-  type CacheStats,
-  type InMemoryCacheOptions,
-  InMemoryCacheProvider,
-  type NamespacedCache,
-  PostgresCacheInvalidator,
-  type PostgresCacheInvalidatorOptions,
-} from "./cache";
-
-// === Observability ===
-
-export {
-  type AlertCondition,
-  type AlertEffect,
-  AlertEngine,
-  type AlertEngineOptions,
-  type AlertEvaluationResult,
-  type AlertHandler,
-  type AlertOperator,
-  type AlertSeverity,
-  defineSystemAlert,
-  type SystemAlertDefinition,
-} from "./observability/alert-engine";
-export { consoleLogger } from "./observability/console-logger";
-export { InMemoryExecutionLogger } from "./observability/execution-logger";
-export type {
-  Counter,
-  Histogram,
-  InstrumentOptions,
-  Meter,
-  MetricAttributes,
-  MetricAttributeValue,
-} from "./observability/meter";
-export { NoopMeter, noopMeter } from "./observability/meter";
-export {
-  InMemoryMetricsCollector,
-  type MetricSnapshot,
-  type MetricsCollector,
-  type MetricsSummary,
-  noopMetricsCollector,
-} from "./observability/metrics";
-export type { Observability } from "./observability/observability-registry";
-export {
-  getObservability,
-  resetObservability,
-  setObservability,
-} from "./observability/observability-registry";
-export {
-  createStructuredLogger,
-  createTestLogSink,
-  type LogLevel,
-  type LogSink,
-  type StructuredLogEntry,
-  type StructuredLoggerOptions,
-} from "./observability/structured-logger";
-export {
-  getCurrentTrace,
-  getTraceDepth,
-  type TraceState,
-  withTrace,
-} from "./observability/trace-context";
-export type {
-  Span,
-  SpanAttributes,
-  SpanAttributeValue,
-  SpanKind,
-  SpanStatus,
-  SpanStatusCode,
-  StartSpanOptions,
-  Tracer,
-} from "./observability/tracer";
-export { NoopTracer, noopTracer } from "./observability/tracer";
-
-// === AI service ===
-
-export { createNoopAIService } from "./ai/ai-service";
-
-// === AI Boundary (server-only — heavyweight runtime classes) ===
-// AI Pattern Detection: concrete `PatternDetector` moved to
-// @linchkit/cap-ai-provider (Spec 56 Phase 2 Step 2c). Core re-exports
-// only the shared `PatternInsight` contract that
-// ProposalEngine.createFromInsight() consumes.
-
-export type { PatternEvidence, PatternInsight, PatternType } from "./ai";
-export { AIBoundary, AIBoundaryError } from "./ai";
-
-// === AI Audit (server-only — compliance audit trail) ===
-
-export {
-  type AIAuditEntry,
-  type AIAuditEventType,
-  AIAuditLogger,
-  type AIAuditLoggerOptions,
-  type AIAuditQueryOptions,
-  type AIAuditRiskLevel,
-  type IntentResolutionAuditPayload,
-} from "./ai";
-
-// === AI Prompt Sanitizer (server-only — injection detection + PII redaction) ===
-
-export {
-  detectInjection,
-  type InjectionDetectionConfig,
-  type InjectionDetectionResult,
-  type InjectionPattern,
-  type PIIPattern,
-  type PIISanitizationConfig,
-  type PIISanitizationResult,
-  type PIIType,
-  type PromptSanitizerOptions,
-  type SanitizationResult,
-  sanitizePII,
-  sanitizePrompt,
-  sanitizeRecordForAI,
-} from "./ai";
-
-// === AI Intent Resolver (server-only — natural language → discriminated Intent, Spec 52 §2.2 / §2.5) ===
-
-export {
-  ALTERNATIVES_CONFIDENCE_THRESHOLD as INTENT_ALTERNATIVES_CONFIDENCE_THRESHOLD,
-  buildIntentSystemPrompt,
-  DEFAULT_MAX_HISTORY_MESSAGES as INTENT_DEFAULT_MAX_HISTORY_MESSAGES,
-  type Intent,
-  type IntentAlternative,
-  type IntentCatalogEntry,
-  type IntentClarification,
-  type IntentHistoryMessage,
-  type IntentMatch,
-  type IntentMultiStep,
-  type IntentNoMatch,
-  type IntentOntology,
-  type IntentPromptOptions,
-  type IntentResolverOptions,
-  type IntentSlot,
-  type IntentStep,
-  MAX_ALTERNATIVES as INTENT_MAX_ALTERNATIVES,
-  MIN_CONFIDENCE as INTENT_MIN_CONFIDENCE,
-  type ResolveIntentDeps,
-  type ResolveIntentInput,
-  resolveIntent,
-} from "./ai";
-
-// === AI Output Validator (server-only — output safety checks) ===
-
-export {
-  type OutputValidationResult,
-  type OutputValidationRule,
-  type OutputValidatorConfig,
-  type OutputViolation,
-  type OutputViolationSeverity,
-  type OutputViolationType,
-  sanitizeAIOutput,
-  validateAIOutput,
-} from "./ai";
-
-// === AI Proposal Validator (server-only — proposal security checks) ===
-
-export {
-  buildCompatibilitySnapshot,
-  type CompatibilityChange,
-  type CompatibilityIssue,
-  type CompatibilityRegistrySnapshot,
-  type CompatibilityResult,
-  type CompatibilitySeverity,
-  compatibilityCheck,
-  createExtendedProposalValidator,
-  createProposalValidator,
-  type DryRunModelError,
-  type DryRunResult,
-  type DryRunSideEffects,
-  dryRunProposal,
-  type EntityCreateChange,
-  type EntityDeleteChange,
-  type EntityReference,
-  type EntityRenameChange,
-  type EnumOptionsChange,
-  type ExtendedPhaseStatus,
-  type ExtendedPhaseSummary,
-  type ExtendedValidationResult,
-  type ExtendedValidatorConfig,
-  type ExtendedValidatorInput,
-  type FieldAddChange,
-  type FieldConstraintChange,
-  type FieldDropChange,
-  type FieldTypeChange,
-  type ProposalChange,
-  type ProposalChangeType,
-  type ProposalCustomRule,
-  type ProposalRiskLevel,
-  type ProposalValidationResult,
-  type ProposalValidatorConfig,
-  type ProposalViolation,
-  validateProposal as validateAIProposal,
-  validateProposalExtended,
-} from "./ai";
-
-// === AI Anomaly Detector ===
-// Moved to @linchkit/cap-ai-provider (Spec 56 Phase 2 Step 2c).
-
-// === Security: data masking (server-only — uses node:crypto) ===
-
-export {
-  canUnmask,
-  type MaskRecordOptions,
-  maskRecord,
-  maskRecords,
-  maskValue,
-  resolveFieldMasking,
-} from "./security";
-
-// === Flow engine ===
-
-export {
-  createFlowRegistry,
-  createFlowStepContext,
-  createSyncFlowEngine,
-  createTriggerBinding,
-  type FlowEngine,
-  type FlowExecuteActionOptions,
-  type FlowRegistry,
-  FlowRegistryImpl,
-  type FlowStepContext,
-  type FlowStepContextDeps,
-  type TriggerBinding,
-} from "./flow";
-
-// === Persistence: database, Drizzle ORM, system tables ===
-
-export type { OverlayChangeListener, OverlayRegistry } from "./overlay/overlay-registry";
-export { DefaultOverlayRegistry } from "./overlay/overlay-registry";
-export type { PromotionPlan } from "./overlay/promote";
-export { generateFieldCode, generateMigrationSql, generatePromotionPlan } from "./overlay/promote";
-export {
-  checkConnection,
-  closeDatabase,
-  createDatabase,
-  type DatabaseConfig,
-} from "./persistence/database";
-export { DrizzleApprovalStore } from "./persistence/drizzle-approval-store";
-export { DrizzleConfigStore } from "./persistence/drizzle-config-store";
-export { DrizzleDataProvider, type I18nQueryOptions } from "./persistence/drizzle-data-provider";
-export { DrizzleExecutionLogger } from "./persistence/drizzle-execution-logger";
-export { DrizzleOverlayStore } from "./persistence/drizzle-overlay-store";
-export * as drizzleSchema from "./persistence/drizzle-schema";
-export { DrizzleTransactionManager } from "./persistence/drizzle-transaction-manager";
-export { InMemoryOverlayStore } from "./persistence/in-memory-overlay-store";
-export { type FindManyOptions, InMemoryStore } from "./persistence/in-memory-store";
-export { OverlayAwareDataProvider } from "./persistence/overlay-aware-data-provider";
-export {
-  fieldOverlaysTable,
-  overlayStatusEnum,
-} from "./persistence/overlay-table";
-export {
-  approvalStatusEnum,
-  approvalsTable,
-  configScopeEnum,
-  configTable,
-  configVersionsTable,
-  eventStatusEnum,
-  eventsTable,
-  executionStatusEnum,
-  executionsTable,
-  linchkitSchema,
-  overrideTargetTypeEnum,
-  tenantOverridesTable,
-} from "./persistence/system-tables";
-export { TableRegistry } from "./persistence/table-registry";
-export {
-  type OverrideTargetType,
-  type TenantOverride,
-  TenantOverrideStore,
-} from "./persistence/tenant-override-store";
-
-// === Security: tenant isolation ===
-
-export {
-  createTenantAwareDataProvider,
-  createTenantIsolationMiddleware,
-  defaultTenantResolver,
-  type TenantIsolationMiddlewareOptions,
-  type TenantResolver,
-} from "./security/tenant-isolation";
-
-// === Ontology: unified semantic facade ===
-
-export {
-  createOntologyRegistry,
-  type EntityDescriptor,
-  type OntologyRegistry,
-  type OntologyRegistryDeps,
-  type RelationDescriptor,
-} from "./ontology";
-
-// === Life-system: Sense layer (Spec 55) ===
-
-export type {
-  AwarenessEngineOptions,
-  EvolutionRuntime,
-  EvolutionRuntimeOptions,
-  SensorDefinitionConfig,
-  SignalBus,
-  SignalBusOptions,
-  SignalHandler,
-} from "./life-system";
-export {
-  createAttentionBudget,
-  createAwarenessEngine,
-  createDispatchQuery,
-  createEvolutionRuntime,
-  createSignalBus,
-  createUsageImportanceGraph,
-  defineSensor,
-} from "./life-system";
-
-// === Deployment: health checks, graceful shutdown, environment ===
-
-export {
-  type AggregatedHealthStatus,
-  createCacheCheck,
-  createDatabaseCheck,
-  createEntityCheck,
-  createEventBusCheck,
-  type DeployEvent,
-  type DeployWebhookConfig,
-  DeployWebhookHandler,
-  detectEnvironment,
-  type EnvironmentConfig,
-  type EnvironmentFeatureFlags,
-  type EnvironmentName,
-  GracefulShutdownManager,
-  type GracefulShutdownManagerOptions,
-  type HealthCheckFn,
-  HealthCheckRegistry,
-  type HealthCheckRegistryOptions,
-  type HealthCheckResult,
-  type HealthStatus,
-  livenessCheck,
-  type ShutdownHook,
-  type ShutdownPhase,
-  type ShutdownStatus,
-  validateRequiredEnvVars,
-  type WebhookHandleResult,
-} from "./deployment";
-
-// === Doctor — project health check registry + built-in checks ===
-
-export type {
-  CheckCategory,
-  CheckStatus,
-  DoctorCheck,
-  DoctorCheckResult,
-  DoctorContext,
-} from "./doctor";
-export { builtinChecks, clearDoctorChecks, getDoctorChecks, registerDoctorCheck } from "./doctor";
-
-// === Addon discovery (Spec 57) ===
-
-export { scanAddonsPath } from "./capability/addon-scanner";
-export { resolveAutoInstall } from "./capability/auto-install";
+export * from "./exports/server/addon-discovery";
+export * from "./exports/server/ai";
+export * from "./exports/server/automation";
+export * from "./exports/server/cache";
+export * from "./exports/server/deployment";
+export * from "./exports/server/doctor";
+export * from "./exports/server/engines";
+export * from "./exports/server/entity";
+export * from "./exports/server/event";
+export * from "./exports/server/flow";
+export * from "./exports/server/life-system";
+export * from "./exports/server/observability";
+export * from "./exports/server/ontology";
+export * from "./exports/server/persistence";
+export * from "./exports/server/security";
